+++ include/vsmc/math/vmath.hpp
@@ -34,6 +34,7 @@
 
 #include <vsmc/internal/config.h>
 #include <vsmc/math/constants.hpp>
+#include <algorithm>
 #include <cmath>
 
 #if VSMC_USE_MKL_VML
@@ -565,6 +566,102 @@ VSMC_DEFINE_MATH_VMATH_1(std::tgamma, tgamma)
 
 /// @}
 // vSpecial
+
+/// \defgroup vPackUnpack Pack and unpack vectors
+/// \brief Pack and unpack vectors into/from contiguous storage
+/// \ingroup vMath
+/// @{
+
+/// \brief Pack vector `src` into vector `dst`, which has a stride `1`, such
+/// that `dst[i] = src[i * stride]`.
+template <typename RandomIter, typename IntType, typename OutputIter>
+inline void pack_s(
+    std::size_t n, RandomIter src, IntType stride, OutputIter dst)
+{
+    if (stride == 1)
+        std::copy_n(src, n, dst);
+    for (std::size_t i = 0; i != n; ++i, src += stride, ++dst) {
+        *dst =
+            static_cast<typename std::iterator_traits<OutputIter>::value_type>(
+                *src);
+    }
+}
+
+/// \brief Pack vector `src` into vector `dst`, which has a stride `1`, such
+/// that `dst[i] = src[index[i]]`.
+template <typename RandomIter, typename InputIter, typename OutputIter>
+inline void pack_i(
+    std::size_t n, RandomIter src, InputIter index, OutputIter dst)
+{
+    for (std::size_t i = 0; i != n; ++i, ++index, ++dst) {
+        *dst =
+            static_cast<typename std::iterator_traits<OutputIter>::value_type>(
+                src[static_cast<typename std::iterator_traits<
+                    RandomIter>::difference_type>(*index)]);
+    }
+}
+
+/// \brief Pack vector `src` into vector `dst`, which has a stride `1`, such
+/// that only `src[i]` with `mask[i]` being `true` (when converted to `bool`)
+/// are copied.
+template <typename InputIterSrc, typename InputIterMask, typename OutputIter>
+inline void pack_m(
+    std::size_t n, InputIterSrc src, InputIterMask mask, OutputIter dst)
+{
+    for (std::size_t i = 0; i != n; ++i, ++src, ++mask) {
+        if (static_cast<bool>(*mask)) {
+            *dst++ = static_cast<
+                typename std::iterator_traits<OutputIter>::value_type>(*src);
+        }
+    }
+}
+
+/// \brief Unpack a vector `src`, which has a stride `1`, into a vector `dst`,
+/// such that `dst[i * stride] = src[i]`.
+template <typename InputIter, typename IntType, typename RandomIter>
+inline void unpack_s(
+    std::size_t n, InputIter src, IntType stride, RandomIter dst)
+{
+    if (stride == 1)
+        std::copy_n(src, n, dst);
+    for (std::size_t i = 0; i != n; ++i, ++src, dst += stride) {
+        *dst =
+            static_cast<typename std::iterator_traits<RandomIter>::value_type>(
+                *src);
+    }
+}
+
+/// \brief Unpack a vector `src`, which has a stride `1`, into a vector `dst`,
+/// such that `dst[index[i]] = src[i]`.
+template <typename InputIterSrc, typename InputIterIndex, typename RandomIter>
+inline void unpack_i(
+    std::size_t n, InputIterSrc src, InputIterIndex index, RandomIter dst)
+{
+    for (std::size_t i = 0; i != n; ++i, ++src, ++index) {
+        dst[static_cast<typename std::iterator_traits<
+            RandomIter>::difference_type>(*index)] =
+            static_cast<typename std::iterator_traits<RandomIter>::value_type>(
+                *src);
+    }
+}
+
+/// \brief Unpack a vector `src`, which has a stride `1`, into a vector `dst`,
+/// such that only `dst[i]` with `mask[i]` being `true` (when converted to
+/// `bool`) is modified.
+template <typename InputIterSrc, typename InputIterMask, typename OutputIter>
+inline void unpack_m(
+    std::size_t n, InputIterSrc src, InputIterMask mask, OutputIter dst)
+{
+    for (std::size_t i = 0; i != n; ++i, ++mask, ++dst) {
+        if (static_cast<bool>(*mask)) {
+            *dst = static_cast<
+                typename std::iterator_traits<OutputIter>::value_type>(*src++);
+        }
+    }
+}
+
+/// @}
+// vPackUnpack
 
 } // namespace vsmc